+++ document/changelog.md
@@ -2,6 +2,7 @@ unreleased
 
 * rework API
 * interpret unknown xref stream entry type as reference to null object
+* support crypto handler version 4 (V2 and AESV2)
 
 0.0.4.0